fix: 내역 삭제 시 id undefined라 삭제되지 않는 문제 수정
