+++ components/monthlyInfo.js
@@ -1,6 +1,6 @@
 import { totalIncomeData, totalExpenseData } from "../utils/transaction.js";
 import { transactionStore } from "../store/index.js";
-import { setFilteringState } from "../pages.js";
+import { setFilteringState } from "../pages/mainPage.js";
 import { renderTransactionList } from "./transactionsList.js";
 import { formatMoney } from "../utils/format.js";
 import { dateStore } from "../store/index.js";
