+++ pages/chartPage.js
@@ -1,0 +1,14 @@
+export function createChartPage() {
+  return `
+    <div class="chart-page">
+      <h2>차트</h2>
+    </div>
+  `;
+}
+
+export function renderChartPage() {
+  const mainContainer = document.getElementById("main-container");
+  if (mainContainer) {
+    mainContainer.innerHTML = createChartPage();
+  }
+}
