+++ pages/mainPage.js
@@ -1,13 +1,12 @@
 import {
   renderMonthlyInfo,
   renderTotalCount,
-} from "./components/monthlyInfo.js";
-import { renderInputBar } from "./components/inputBar.js";
-import { renderTransactionList } from "./components/transactionsList.js";
-import { totalIncomeData, totalExpenseData } from "./utils/transaction.js";
-import { transactionStore } from "./store/index.js";
-import { renderCalendar, renderCalendarInfo } from "./components/calendar.js";
-import { dateStore } from "./store/index.js";
+} from "../components/monthlyInfo.js";
+import { renderInputBar } from "../components/inputBar.js";
+import { renderTransactionList } from "../components/transactionsList.js";
+import { totalIncomeData, totalExpenseData } from "../utils/transaction.js";
+import { transactionStore } from "../store/index.js";
+import { dateStore } from "../store/index.js";
 
 let isIncomeChecked = true;
 let isExpenseChecked = true;
@@ -27,23 +26,6 @@ export function createMainPage() {
     <div class="container main-page">
       <div id="monthly-info-container"></div>
       <div id="transaction-list-container"></div>
-    </div>
-  `;
-}
-
-export function createCalendarPage() {
-  return `
-    <div class="container calendar-page">
-      <div class="calendar-container"></div>
-      <div class="calendar-info-container"></div>
-    </div>
-  `;
-}
-
-export function createChartPage() {
-  return `
-    <div class="chart-page">
-      <h2>차트</h2>
     </div>
   `;
 }
@@ -90,29 +72,4 @@ export function renderMainPage() {
   }
 
   renderTransactionList(isIncomeChecked, isExpenseChecked);
-}
-
-export function renderCalendarPage() {
-  const mainContainer = document.getElementById("main-container");
-  if (mainContainer) {
-    mainContainer.innerHTML = createCalendarPage();
-  }
-
-  const calendarContainer = mainContainer.querySelector(".calendar-container");
-  if (calendarContainer) {
-    renderCalendar(calendarContainer);
-  }
-  const calendarInfoContainer = mainContainer.querySelector(
-    ".calendar-info-container"
-  );
-  if (calendarInfoContainer) {
-    renderCalendarInfo(calendarInfoContainer);
-  }
-}
-
-export function renderChartPage() {
-  const mainContainer = document.getElementById("main-container");
-  if (mainContainer) {
-    mainContainer.innerHTML = createChartPage();
-  }
 }
